refactor: split classifier logic into dedicated module
