+++ geo_bot.py
@@ -1,8 +1,5 @@
 import os
 
-import numpy as np
-import requests
-import tensorflow as tf
 from telegram import Update
 from telegram.ext import (
     Application,
@@ -11,98 +8,15 @@ from telegram.ext import (
     MessageHandler,
     filters,
 )
-from tensorflow.keras.preprocessing import image
+
+from classifier import TRASHNET_MODEL, classify_trash
 
 # Путь для сохранения загруженных фото
 PHOTO_SAVE_PATH = "./photos"
 
-# Optional external inference API settings
-EXTERNAL_CLASSIFIER_API_URL = os.environ.get("EXTERNAL_CLASSIFIER_API_URL")
-EXTERNAL_CLASSIFIER_API_TOKEN = os.environ.get("EXTERNAL_CLASSIFIER_API_TOKEN")
-EXTERNAL_CLASSIFIER_TIMEOUT = int(os.environ.get("EXTERNAL_CLASSIFIER_TIMEOUT", "20"))
-
 # Убедимся, что директория для фото существует
 if not os.path.exists(PHOTO_SAVE_PATH):
     os.makedirs(PHOTO_SAVE_PATH)
-
-
-def classify_with_external_api(image_path: str) -> str:
-    """Classify image using external HTTP API.
-
-    Expected API response formats:
-    - {"classification": "Plastic"}
-    - {"label": "Plastic"}
-    """
-    if not EXTERNAL_CLASSIFIER_API_URL:
-        raise RuntimeError(
-            "Classification model is unavailable and EXTERNAL_CLASSIFIER_API_URL is not configured."
-        )
-
-    headers = {}
-    if EXTERNAL_CLASSIFIER_API_TOKEN:
-        headers["Authorization"] = f"Bearer {EXTERNAL_CLASSIFIER_API_TOKEN}"
-
-    with open(image_path, "rb") as image_file:
-        response = requests.post(
-            EXTERNAL_CLASSIFIER_API_URL,
-            files={"file": image_file},
-            headers=headers,
-            timeout=EXTERNAL_CLASSIFIER_TIMEOUT,
-        )
-
-    if response.status_code >= 400:
-        raise RuntimeError(
-            f"External classifier API error: {response.status_code}"
-        )
-
-    payload = response.json()
-    classification = payload.get("classification") or payload.get("label")
-    if not classification:
-        raise RuntimeError("External classifier API returned invalid payload.")
-
-    return classification
-
-
-# Загрузка модели (заглушка)
-def load_trashnet_model():
-    """
-    Stub function for loading the model.
-    Returns None because the model is currently unavailable.
-    """
-    print("Model loading is stubbed out as the model file is unavailable.")
-    return None
-
-
-# Загружаем модель один раз при старте
-TRASHNET_MODEL = load_trashnet_model()
-
-
-# Функция для классификации мусора на фотографии
-def classify_trash(image_path, model):
-    """Classifies trash on an image with a loaded model.
-
-    If local model is unavailable, tries EXTERNAL_CLASSIFIER_API_URL.
-    """
-    if model is None:
-        return classify_with_external_api(image_path)
-
-    img = image.load_img(image_path, target_size=(224, 224))
-    img_array = image.img_to_array(img)
-    img_array = np.expand_dims(img_array, axis=0)
-    img_array /= 255.0
-
-    prediction = model.predict(img_array)
-    trash_classes = [
-        "Glass",
-        "Paper",
-        "Cardboard",
-        "Plastic",
-        "Metal",
-        "Trash",
-    ]
-    trash_class = trash_classes[np.argmax(prediction)]
-
-    return trash_class
 
 
 # Обработчик команды /start
@@ -140,7 +54,7 @@ async def photo(update: Update, context: CallbackContext) -> None:
     photo_path = os.path.join(PHOTO_SAVE_PATH, f"{file_id}.jpg")
     await new_file.download_to_drive(photo_path)
 
-    # Классификация мусора на фото (используем уже загруженную модель)
+    # Классификация мусора на фото
     trash_class = classify_trash(photo_path, TRASHNET_MODEL)
 
     # Отправка сообщения с результатом
